Remove DumpWorker module
